+++ docs/examples/example_intersections.py
@@ -7,6 +7,7 @@ from compas.geometry import Point
 from compas_viewer import Viewer
 
 from compas_libigl.intersections import intersection_rays_mesh
+from compas_libigl.intersections import barycenter_to_point
 
 # ==============================================================================
 # Input geometry
@@ -16,7 +17,6 @@ mesh = Mesh.from_obj(compas.get("tubemesh.obj"))
 
 trimesh = mesh.copy()
 trimesh.quads_to_triangles()
-
 # ==============================================================================
 # Rays
 # ==============================================================================
@@ -24,8 +24,8 @@ trimesh.quads_to_triangles()
 base = Point(*mesh.centroid())
 base.z = 0
 
-theta = np.linspace(0, np.pi, 20, endpoint=False)
-phi = np.linspace(0, 2 * np.pi, 20, endpoint=False)
+theta = np.linspace(0, np.pi, 5, endpoint=False)
+phi = np.linspace(0, 2 * np.pi, 5, endpoint=False)
 theta, phi = np.meshgrid(theta, phi)
 theta = theta.ravel()
 phi = phi.ravel()
@@ -38,30 +38,27 @@ xyz = np.vstack((x, y, z)).T
 mask = xyz[:, 2] > 0
 hemi = xyz[mask]
 
+lines = []
 rays = []
 for x, y, z in hemi:
     point = Point(x, y, z)
     vector = point - base
     vector.unitize()
+    lines.append(Line(base, base + vector))
     rays.append((base, vector))
 
 # ==============================================================================
 # Intersections
 # ==============================================================================
 
-index_face = {index: face for index, face in enumerate(mesh.faces())}
+hits_per_rays = intersection_rays_mesh(rays, trimesh.to_vertices_and_faces())
 
-hits_per_ray = intersection_rays_mesh(rays, mesh.to_vertices_and_faces())
-
-intersections = []
-for ray, hits in zip(rays, hits_per_ray):
-    if hits:
-        base, vector = ray
-        index = hits[0][0]
-        distance = hits[0][3]
-        face = index_face[index]
-        point = base + vector * distance
-        intersections.append(point)
+intersection_points = []
+for hits_per_ray in hits_per_rays:
+    if hits_per_ray:
+        for hit in hits_per_ray:
+            pt, idx, u, v, w = hit
+            intersection_points.append(pt)
 
 # ==============================================================================
 # Visualisation
@@ -71,7 +68,12 @@ viewer = Viewer(width=1600, height=900)
 
 viewer.scene.add(mesh, opacity=0.7, show_points=False)
 
-for intersection in intersections:
-    viewer.scene.add(Line(base, intersection), linecolor=Color.blue(), linewidth=3)
+for point in intersection_points:
+    viewer.scene.add(Line(base, point), linecolor=Color.blue(), linewidth=3)
+    viewer.scene.add(point, pointcolor=Color.red(), pointsize=10)
+
+for line in lines:
+    for i in range(20):
+        viewer.scene.add(line.point_at(i / 20), pointcolor=Color.red(), pointsize=5)
 
 viewer.show()