+++ docs/examples/example_mapping_patterns.py
@@ -9,13 +9,11 @@ from compas_viewer.config import Config
 from compas_libigl.mapping import map_pattern_to_mesh
 
 
-
 # ==============================================================================
 # Input geometry: 3D Mesh
 # ==============================================================================
 
 mesh = Mesh.from_obj(Path(__file__).parent.parent.parent / "data" / "minimal_surface.obj")
-
 
 
 for vertex in mesh.vertices():
@@ -32,7 +30,7 @@ fixed_vertices = []
 
 for vertex in mesh.vertices():
     x, y, z = mesh.vertex_attributes(vertex, "xyz")  # type: ignore
-    if abs(z-aabb.zmin) < 1e-3 or abs(z-aabb.zmax) < 1e-3:
+    if abs(z - aabb.zmin) < 1e-3 or abs(z - aabb.zmax) < 1e-3:
         fixed_vertices.append(vertex)
 
 # ==============================================================================